+++ views/blocks/header/header.js
@@ -9,8 +9,6 @@ require('../../styles/tabs/tabs');
 require('../../styles/avatar/avatar');
 require('./head_arrow.png');
 
-const formValidation = require('../../../libs/clientScripts/form-validation');
-
 const search = document.getElementById('search');
 search.addEventListener('submit', function (event) {
     event.preventDefault();
@@ -38,3 +36,32 @@ $('.tabs__link').on('click', function () {
 });
 
 formValidation(success, '.auth-form');
+
+function formValidation(success, formClassName) {
+    $(formClassName).on('submit', function () {
+        const form = this;
+        const msg = $(form).serialize();
+        const $formMessage = $(`${formClassName}-message`);
+        $formMessage.html('');
+
+        const $btnPrimary = $(form).find('.btn_primary');
+        $btnPrimary.prop('disabled', true);
+
+        $.ajax({
+            type: 'POST',
+            url: form.action,
+            data: msg,
+
+            success: success.bind(null, form, $formMessage, $btnPrimary),
+
+            error: function (res) {
+                $formMessage.html(res.responseText).addClass('error').removeClass('success');
+                $(form).find('input[name="password"]').val('');
+                $btnPrimary.prop('disabled', false);
+            }
+        });
+
+        return false;
+    });
+}
+
